Divide the code into modules
